requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
drf-yasg==1.21.7
environs==9.5.0
inflection==0.5.1
marshmallow==3.20.1
packaging==23.1
psycopg2==2.9.7
Pygments==2.16.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
